+++ examples/resources/neon_branch/resource.tf
@@ -1,0 +1,8 @@
+resource "neon_project" "example" {
+  name = "foo"
+}
+
+resource "neon_branch" "example" {
+  project_id = neon_project.example.id
+  name       = "foo"
+}