+++ internal/provider/helper.go
@@ -1,0 +1,72 @@
+package provider
+
+import (
+	"context"
+	"errors"
+	"net/http"
+	"time"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	neon "github.com/kislerdm/neon-sdk-go"
+)
+
+func pgSettingsToMap(v neon.PgSettingsData) map[string]interface{} {
+	o := make(map[string]interface{}, len(v))
+	for k, v := range v {
+		o[k] = v
+	}
+	return o
+}
+
+func mapToPgSettings(v map[string]interface{}) neon.PgSettingsData {
+	o := make(neon.PgSettingsData, len(v))
+	for k, v := range v {
+		o[k] = v
+	}
+	return o
+}
+
+func intValidationNotNegative(v interface{}, s string) (warn []string, errs []error) {
+	if v.(int) < 0 {
+		errs = append(errs, errors.New(s+" must be not negative"))
+		return
+	}
+	return
+}
+
+type delay struct {
+	delay  time.Duration
+	maxCnt uint8
+}
+
+func (r *delay) Retry(
+	fn func(context.Context, *schema.ResourceData, interface{}) error,
+	ctx context.Context, d *schema.ResourceData, meta interface{},
+) diag.Diagnostics {
+	var i uint8
+	var err error
+	for i < r.maxCnt {
+		switch e := fn(ctx, d, meta).(type) {
+		case nil:
+			return nil
+		case neon.Error:
+			switch e.HTTPCode {
+			case 200:
+				return nil
+			case http.StatusTooManyRequests, http.StatusInternalServerError:
+				err = e
+				i++
+				time.Sleep(r.delay)
+			}
+		default:
+			return diag.FromErr(e)
+		}
+	}
+	return diag.FromErr(err)
+}
+
+var projectReadiness = delay{
+	delay:  5 * time.Second,
+	maxCnt: 120,
+}