+++ CHANGELOG.md
@@ -5,7 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1]
+## [0.0.2] - unreleased
+
+### Added
+
+- Branch resource:
+
+```terraform
+resource "neon_project" "this" {
+  name = "foo"
+}
+
+resource "neon_branch" "this" {
+  project_id = neon_project.foo.id
+  name       = "bar"
+}
+```
+
+- Backoff+retry mechanism: operation is retried after the delay of 5 sec. API response's HTTP codes are 500, or 429.
+  Total number of attempts is limited to 120 per operation.
+
+### Changed
+
+- Bumped [Neon Go SDK](https://pkg.go.dev/github.com/kislerdm/neon-sdk-go) v0.1.3
+- Added errors handling for the project resource
+
+## [0.0.1] - 2022-01-05
 
 ### Added
 
@@ -15,7 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 terraform {
   required_providers {
     neon = {
-      source = "hashicorp.com/kislerdm/neon"
+      source = "kislerdm/neon"
     }
   }
 }
@@ -26,5 +51,7 @@ provider "neon" {}
 - Project resource:
 
 ```terraform
-resource "neon_project" "foo" {}
+resource "neon_project" "this" {
+  name = "foo"
+}
 ```